requirements.txt
beautifulsoup4==4.13.3
chromadb==0.4.22
Flask==3.1.0
Flask_Cors==4.0.1
flask_wtf==1.2.2
langchain==0.3.18
langchain_community==0.3.17
langchain_core==0.3.34
mistralai==1.5.0
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
